# backend/openapi.yml
openapi: 3.0.3
info:
  title: TekDocs API
  version: 0.1.9
  description: Self-hosted MSP knowledge and inventory API
paths:
...
  /api/v1/people:
    get:
      operationId: people_msp_list
      parameters:
      - in: query
        name: filter_field
        schema:
          type: string
      - in: query
        name: filter_value
        schema:
          type: string
      - in: query
        name: ordering
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
        description: Search every displayed person and association field.
      tags:
      - people
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeopleResult'
          description: ''
        '403':
          description: No response body
    post:
      operationId: people_msp_create
      tags:
      - people
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PersonWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PersonWrite'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '403':
          description: No response body
  /api/v1/people/{person_entity_id}:
    get:
      operationId: people_msp_retrieve
      parameters:
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - people
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '404':
          description: No response body
    patch:
      operationId: people_msp_update
      parameters:
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - people
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '404':
          description: No response body
    delete:
      operationId: people_msp_archive
      parameters:
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - people
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
        '404':
          description: No response body
...
  /api/v1/workspaces/organizations/{organization_entity_id}/people:
    get:
      operationId: people_organization_list
      parameters:
      - in: query
        name: filter_field
        schema:
          type: string
      - in: query
        name: filter_value
        schema:
          type: string
      - in: query
        name: ordering
        schema:
          type: string
      - in: path
        name: organization_entity_id
        schema:
          type: string
          format: uuid
        required: true
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
        description: Search every displayed person and association field.
      tags:
      - workspaces
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeopleResult'
          description: ''
        '404':
          description: No response body
    post:
      operationId: people_organization_create
      parameters:
      - in: path
        name: organization_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PersonWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PersonWrite'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '404':
          description: No response body
  /api/v1/workspaces/organizations/{organization_entity_id}/people/{person_entity_id}:
    get:
      operationId: people_organization_retrieve
      parameters:
      - in: path
        name: organization_entity_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '404':
          description: No response body
    patch:
      operationId: people_organization_update
      parameters:
      - in: path
        name: organization_entity_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPersonWrite'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          description: ''
        '404':
          description: No response body
    delete:
      operationId: people_organization_archive
      parameters:
      - in: path
        name: organization_entity_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: person_entity_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
        '404':
          description: No response body
...
components:
  schemas:
...
    PatchedPersonWrite:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 240
          minLength: 1
        preferred_name:
          type: string
          maxLength: 160
        kind:
          $ref: '#/components/schemas/PersonAssociationKindEnum'
        role:
          type: string
          maxLength: 160
        responsibility:
          type: string
          maxLength: 240
        location:
          type: string
          maxLength: 160
        office:
          type: string
          maxLength: 120
        site_id:
          type: string
          format: uuid
          nullable: true
        structured_location_id:
          type: string
          format: uuid
          nullable: true
        phone:
          type: string
          maxLength: 64
        email:
          type: string
          format: email
          maxLength: 254
...
    PeopleResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        page:
          type: integer
        page_size:
          type: integer
        count:
          type: integer
        has_more:
          type: boolean
      required:
      - count
      - has_more
      - page
      - page_size
      - results
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        association_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
          nullable: true
        full_name:
          type: string
        preferred_name:
          type: string
        kind:
          $ref: '#/components/schemas/PersonAssociationKindEnum'
        role:
          type: string
        responsibility:
          type: string
        location:
          type: string
        office:
          type: string
        site_id:
          type: string
          format: uuid
          nullable: true
        structured_location_id:
          type: string
          format: uuid
          nullable: true
        phone:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - association_id
      - created_at
      - email
      - full_name
      - id
      - kind
      - location
      - office
      - organization_id
      - phone
      - preferred_name
      - responsibility
      - role
      - site_id
      - structured_location_id
      - updated_at
    PersonAssociationKindEnum:
      enum:
      - employee
      - contact
      type: string
      description: |-
        * `employee` - Employee
        * `contact` - Contact
    PersonWrite:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 240
          minLength: 1
        preferred_name:
          type: string
          maxLength: 160
        kind:
          $ref: '#/components/schemas/PersonAssociationKindEnum'
        role:
          type: string
          maxLength: 160
        responsibility:
          type: string
          maxLength: 240
        location:
          type: string
          maxLength: 160
        office:
          type: string
          maxLength: 120
        site_id:
          type: string
          format: uuid
          nullable: true
        structured_location_id:
          type: string
          format: uuid
          nullable: true
        phone:
          type: string
          maxLength: 64
        email:
          type: string
          format: email
          maxLength: 254
      required:
      - full_name
      - kind
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid